requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
Jinja2==2.11.2
pandas==1.4.4
pymongo==4.3.3
mysql-connector-python